Add instruction view
