+++ Cargo.toml
@@ -39,6 +39,7 @@ members = [
     "inflation",
     "instruction",
     "instruction-error",
+    "instruction-view",
     "instructions-sysvar",
     "keccak-hasher",
     "keypair",
@@ -254,6 +255,7 @@ solana-hash = { path = "hash", version = "4.0.0", default-features = false }
 solana-inflation = { path = "inflation", version = "3.0.0" }
 solana-instruction = { path = "instruction", version = "3.0.0", default-features = false }
 solana-instruction-error = { path = "instruction-error", version = "2.0.0" }
+solana-instruction-view = { path = "instruction-view", version = "0.0.0" }
 solana-instructions-sysvar = { path = "instructions-sysvar", version = "3.0.0" }
 solana-keccak-hasher = { path = "keccak-hasher", version = "3.0.0" }
 solana-keypair = { path = "keypair", version = "3.0.0" }
