+++ instruction-view/Cargo.toml
@@ -1,0 +1,25 @@
+[package]
+name = "solana-instruction-view"
+description = "Lightweight instruction representation"
+documentation = "https://docs.rs/solana-instruction-view"
+version = "0.0.0"
+rust-version = "1.81.0"
+authors = { workspace = true }
+repository = { workspace = true }
+homepage = { workspace = true }
+license = { workspace = true }
+edition = { workspace = true }
+
+[features]
+cpi = ["dep:solana-define-syscall", "dep:solana-program-error"]
+
+[dependencies]
+solana-account-view = { workspace = true }
+solana-address = { workspace = true }
+solana-program-error = { workspace = true, optional = true }
+
+[target.'cfg(target_os = "solana")'.dependencies]
+solana-define-syscall = { workspace = true, optional = true }
+
+[lints]
+workspace = true
